add the function about convert agent to tools
