+++ src/tools/__init__.py
@@ -1,4 +1,4 @@
-from src.tools.tools import Tool, ToolResult, AsyncTool
+from src.tools.tools import Tool, ToolResult, AsyncTool, make_tool_instance
 from src.tools.deep_analyzer import DeepAnalyzerTool
 from src.tools.deep_researcher import DeepResearcherTool
 from src.tools.python_interpreter import PythonInterpreterTool
@@ -15,4 +15,5 @@ __all__ = [
     "PythonInterpreterTool",
     "AutoBrowserUseTool",
     "PlanningTool",
+    "make_tool_instance",
 ]
